Split up keyword elaboration tests
