+++ src/language/compiling/semantics/keyword-handlers/todo-handler.test.ts
@@ -1,0 +1,13 @@
+import { elaborationSuite, success } from '../test-utilities.test.js'
+
+elaborationSuite('@todo', [
+  [{ 0: '@todo', 1: 'blah' }, success({})],
+  [{ 0: '@todo', 1: { 0: '@@blah' } }, success({})],
+  [
+    {
+      key1: { 0: '@todo', 1: 'this should be replaced with an empty object' },
+      key2: { 0: '@todo' },
+    },
+    success({ key1: {}, key2: {} }),
+  ],
+])
